Move logic related to replication rules to separate module
