+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.11.0] - 2023-09-25
 
+- Move logic related to replication rules to `replicon_core::replication_rules` module.
+
 ### Changed
 
 - Serialize all components and events using varint.
